+++ src/db_iterator.rs
@@ -35,7 +35,11 @@ pub type DBRawIterator<'a> = DBRawIteratorWithThreadMode<'a, DB>;
 /// ```
 /// use rocksdb::{DB, Options};
 ///
-/// let path = "_path_for_rocksdb_storage4";
+/// let tempdir = tempfile::Builder::new()
+///     .prefix("_path_for_rocksdb_storage4")
+///     .tempdir()
+///     .expect("Failed to create temporary path for the _path_for_rocksdb_storage4.");
+/// let path = tempdir.path();
 /// {
 ///     let db = DB::open_default(path).unwrap();
 ///     let mut iter = db.raw_iterator();
@@ -145,7 +149,11 @@ impl<'a, D: DBAccess> DBRawIteratorWithThreadMode<'a, D> {
     /// ```rust
     /// use rocksdb::{DB, Options};
     ///
-    /// let path = "_path_for_rocksdb_storage5";
+    /// let tempdir = tempfile::Builder::new()
+    ///     .prefix("_path_for_rocksdb_storage5")
+    ///     .tempdir()
+    ///     .expect("Failed to create temporary path for the _path_for_rocksdb_storage5.");
+    /// let path = tempdir.path();
     /// {
     ///     let db = DB::open_default(path).unwrap();
     ///     let mut iter = db.raw_iterator();
@@ -182,7 +190,11 @@ impl<'a, D: DBAccess> DBRawIteratorWithThreadMode<'a, D> {
     /// ```rust
     /// use rocksdb::{DB, Options};
     ///
-    /// let path = "_path_for_rocksdb_storage6";
+    /// let tempdir = tempfile::Builder::new()
+    ///     .prefix("_path_for_rocksdb_storage6")
+    ///     .tempdir()
+    ///     .expect("Failed to create temporary path for the _path_for_rocksdb_storage6.");
+    /// let path = tempdir.path();
     /// {
     ///     let db = DB::open_default(path).unwrap();
     ///     let mut iter = db.raw_iterator();
@@ -222,7 +234,11 @@ impl<'a, D: DBAccess> DBRawIteratorWithThreadMode<'a, D> {
     /// ```rust
     /// use rocksdb::{DB, Options};
     ///
-    /// let path = "_path_for_rocksdb_storage7";
+    /// let tempdir = tempfile::Builder::new()
+    ///     .prefix("_path_for_rocksdb_storage7")
+    ///     .tempdir()
+    ///     .expect("Failed to create temporary path for the _path_for_rocksdb_storage7.");
+    /// let path = tempdir.path();
     /// {
     ///     let db = DB::open_default(path).unwrap();
     ///     let mut iter = db.raw_iterator();
@@ -261,7 +277,11 @@ impl<'a, D: DBAccess> DBRawIteratorWithThreadMode<'a, D> {
     /// ```rust
     /// use rocksdb::{DB, Options};
     ///
-    /// let path = "_path_for_rocksdb_storage8";
+    /// let tempdir = tempfile::Builder::new()
+    ///     .prefix("_path_for_rocksdb_storage8")
+    ///     .tempdir()
+    ///     .expect("Failed to create temporary path for the _path_for_rocksdb_storage8.");
+    /// let path = tempdir.path();
     /// {
     ///     let db = DB::open_default(path).unwrap();
     ///     let mut iter = db.raw_iterator();
@@ -395,7 +415,11 @@ pub type DBIterator<'a> = DBIteratorWithThreadMode<'a, DB>;
 /// ```
 /// use rocksdb::{DB, Direction, IteratorMode, Options};
 ///
-/// let path = "_path_for_rocksdb_storage2";
+/// let tempdir = tempfile::Builder::new()
+///     .prefix("_path_for_rocksdb_storage2")
+///     .tempdir()
+///     .expect("Failed to create temporary path for the _path_for_rocksdb_storage2.");
+/// let path = tempdir.path();
 /// {
 ///     let db = DB::open_default(path).unwrap();
 ///     let mut iter = db.iterator(IteratorMode::Start); // Always iterates forward