+++ src/transactions/transaction_db.rs
@@ -56,7 +56,11 @@ type DefaultThreadMode = crate::MultiThreaded;
 ///
 /// ```
 /// use rocksdb::{DB, Options, TransactionDB, SingleThreaded};
-/// let path = "_path_for_transaction_db";
+/// let tempdir = tempfile::Builder::new()
+///     .prefix("_path_for_transaction_db")
+///     .tempdir()
+///     .expect("Failed to create temporary path for the _path_for_transaction_db");
+/// let path = tempdir.path();
 /// {
 ///     let db: TransactionDB = TransactionDB::open_default(path).unwrap();
 ///     db.put(b"my key", b"my value").unwrap();
@@ -994,8 +998,11 @@ impl TransactionDB<SingleThreaded> {
 impl TransactionDB<MultiThreaded> {
     /// Creates column family with given name and options.
     pub fn create_cf<N: AsRef<str>>(&self, name: N, opts: &Options) -> Result<(), Error> {
+        // Note that we acquire the cfs lock before inserting: otherwise we might race
+        // another caller who observed the handle as missing.
+        let mut cfs = self.cfs.cfs.write().unwrap();
         let inner = self.create_inner_cf_handle(name.as_ref(), opts)?;
-        self.cfs.cfs.write().unwrap().insert(
+        cfs.insert(
             name.as_ref().to_string(),
             Arc::new(UnboundColumnFamily { inner }),
         );