+++ src/write_batch.rs
@@ -28,7 +28,11 @@ pub type WriteBatch = WriteBatchWithTransaction<false>;
 /// ```
 /// use rocksdb::{DB, Options, WriteBatchWithTransaction};
 ///
-/// let path = "_path_for_rocksdb_storage1";
+/// let tempdir = tempfile::Builder::new()
+///     .prefix("_path_for_rocksdb_storage1")
+///     .tempdir()
+///     .expect("Failed to create temporary path for the _path_for_rocksdb_storage1");
+/// let path = tempdir.path();
 /// {
 ///     let db = DB::open_default(path).unwrap();
 ///     let mut batch = WriteBatchWithTransaction::<false>::default();
@@ -89,6 +93,22 @@ unsafe extern "C" fn writebatch_delete_callback(state: *mut c_void, k: *const c_
 }
 
 impl<const TRANSACTION: bool> WriteBatchWithTransaction<TRANSACTION> {
+    /// Create a new `WriteBatch` without allocating memory.
+    pub fn new() -> Self {
+        Self {
+            inner: unsafe { ffi::rocksdb_writebatch_create() },
+        }
+    }
+
+    /// Creates `WriteBatch` with the specified `capacity` in bytes. Allocates immediately.
+    pub fn with_capacity_bytes(capacity_bytes: usize) -> Self {
+        Self {
+            // zeroes from default constructor
+            // https://github.com/facebook/rocksdb/blob/0f35db55d86ea8699ea936c9e2a4e34c82458d6b/include/rocksdb/write_batch.h#L66
+            inner: unsafe { ffi::rocksdb_writebatch_create_with_params(capacity_bytes, 0, 0, 0) },
+        }
+    }
+
     /// Construct with a reference to a byte array serialized by [`WriteBatch`].
     pub fn from_data(data: &[u8]) -> Self {
         unsafe {
@@ -408,9 +428,7 @@ impl WriteBatchWithTransaction<false> {
 
 impl<const TRANSACTION: bool> Default for WriteBatchWithTransaction<TRANSACTION> {
     fn default() -> Self {
-        Self {
-            inner: unsafe { ffi::rocksdb_writebatch_create() },
-        }
+        Self::new()
     }
 }