+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * Document that `default` column family doesn't inherit open options of db (
   0xdeafbeef)
 * Expose LRU cache options (athre0z)
+* Add `with_capacity_bytes` to `WriteBatch` (0xdeafbeef)
+* Add `set_compaction_pri` to `Options` (0xdeafbeef)
+* Add `lto` feature to enable link-time optimization using `linker-plugin-lto` (0xdeafbeef)
 
 ## [Breaking Changes]
 * Update jemalloc-sys to 0.6.0 (0xdeafbeef)